Do not fully decode reference scripts (#107)
UTxOs can store reference scripts, but the content of these is not strictly validated by a Node. In practice Plutus scripts should be cbor-wrapped, put if they are not, we would fail trying to fully decode the reference script. This can cause issues when decoding wallet UTxOs for example, that may contain invalid reference scripts.

With this change, we do not fully decode reference scripts, just enough to compute the script hash (even if script is invalid).

* Small doc comment update

* Do not fully decode ref scripts

* Fix format of test file