Replace TxBody by TxContext in Tx building (#113)
This is a breaking change of the Tx building API.

When building redeemers for witnesses, now the users have access to a data record called TxContext which is very similar to the Plutus script context. This improves the pre-computable properties, with additional fields like redeemers, which were not available in the Tx body.

* Replace TxBody by TxContext in Tx building

* Expose TxContext type