Also provide expected signers for spending native scripts (#116)
Previously, we only needed to provide the witness source for spending UTxOs at a native script address. Now we also need to provide expected signers.

Otherwise we don’t know who is gonna sign a multisig (all? none? some?). And we need the correct list of signers to correctly estimate the Tx fees.